Fix scheduler trigerring builds when multiple rows are found (#144)
There was an error in scheduling logic causing duplicate builds to be
schedule.
The problem was in `maybeSingle` call that'd produce an error where
multiple rows were returned. This shouldn't happen because we had unique
constraint for package,version,platform,worklet_version, however the
constraint doesn't work for records where worklets_version is NULL. As a
consequence we ended up having duplicated rows which triggered the error
in `maybeSingle` call.

This PR fixes the logic such that even if more than 1 rows are there,
we'd still limit the call to just 1 row and return that the build has
been already scheduled