Fix index name to avoid errors when updating (#164)
This PR updates the database schema to use identifiable unique index
instead of the previous index that would use coalesce to dedupe NULL
values. This is important as we reference the key when inserting and
hence we were using fields the insert query resulted in an error.

## 🎯 Type of Change

Please delete options that are not relevant:
- 🐛 Bug fix (non-breaking change which fixes an issue)