Add script for checking / updating status field in builds table for ready builds (#173)
## 📝 Description

Because of different issues, sometimes we miss updating the scheduled
status either to failed and completed. This PR adds a script that checks
the scheduled rows older than 10 hours and verifies is the corresponding
artifact is available on maven or not. When "--write" flag is provided
we update the status to failed or completed depending on whether the
artifact is available.