[website] Small visual fixes to RN version selector on Supported libs page (#183)
This PR addresses the following issues:
1) Layout of versions blocks in "RN Versions" container was broken and
the versions block were wrapped in weird way + the "show less" was
always right-aligned. We are now using simple layout to display all the
blocks along with the "show more/less" button
2) After changing RN versions, we need to recalculate library version
wrapping. This is because some RN versions may yield less than 3 library
versions for certain libraries while other can have more than 3
versions. We need to update library wrapping when RN version changes and
when we display library versions for that particular RN version.

We are also adding more RN versions to test data such that the described
problems can be noticed in dev mode.